Move layout proto into config.
