+++ lib/config/build.py
@@ -4,34 +4,45 @@ from build.protobuf import proto, protocc
 proto(
     name="drive_proto",
     srcs=["./drive.proto"],
-    deps=["lib+common_proto", "lib/external+fl2_proto", "lib+layout_proto"],
+    deps=["lib+common_proto", "lib/external+fl2_proto", ".+layout_proto"],
 )
-
 protocc(
     name="drive_proto_lib",
     srcs=[".+drive_proto"],
     deps=[
+        ".+layout_proto_lib",
         "lib+common_proto_lib",
         "lib/external+fl2_proto_lib",
-        "lib+layout_proto_lib",
     ],
 )
+
+proto(
+    name="layout_proto",
+    srcs=["./layout.proto"],
+    deps=["lib+common_proto", "lib/external+fl2_proto"],
+)
+protocc(
+    name="layout_proto_lib",
+    srcs=[".+layout_proto"],
+    deps=["lib+common_proto_lib", "lib/external+fl2_proto_lib"],
+)
+
 proto(
     name="proto",
     srcs=["./config.proto"],
     deps=[
-        "lib+common_proto",
-        "lib+layout_proto",
         ".+drive_proto",
-        "lib/external+fl2_proto",
-        "lib/fluxsource+proto",
-        "lib/fluxsink+proto",
-        "lib/vfs+proto",
-        "lib/usb+proto",
-        "lib/encoders+proto",
+        ".+layout_proto",
+        "lib+common_proto",
         "lib/decoders+proto",
+        "lib/encoders+proto",
+        "lib/external+fl2_proto",
+        "lib/fluxsink+proto",
+        "lib/fluxsource+proto",
         "lib/imagereader+proto",
         "lib/imagewriter+proto",
+        "lib/usb+proto",
+        "lib/vfs+proto",
     ],
 )
