+++ lib/vfs/vfs.proto
@@ -1,7 +1,7 @@
 syntax = "proto2";
 
 import "lib/common.proto";
-import "lib/layout.proto";
+import "lib/config/layout.proto";
 
 message AcornDfsProto
 {
